safari-534.53-branch/LayoutTests/inspector/debugger/event-listener-breakpoints-expected.txt

Tests event listener breakpoints.


Debugger was enabled.

Running: testClickBreakpoint
Script execution paused.
Call stack:
    0) testElementClicked (event-listener-breakpoints.html:7)
    1) addListenerAndClick (event-listener-breakpoints.html:16)
    2)  (:1)
Script execution resumed.

Running: testTimerFiredBreakpoint
Script execution paused.
Call stack:
    0) timerFired (event-listener-breakpoints.html:19)
Script execution resumed.
Debugger was disabled.
